+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/domain/model/userProfile/CompactProfile.java
@@ -1,4 +1,4 @@
-package kr.flab.snapnow.domain.user.model.userProfile;
+package kr.flab.snapnow.domain.user.domain.model.userProfile;
 
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
