+++ snapnow-domain/src/main/java/kr/flab/snapnow/domain/user/domain/model/userProfile/CompactProfileWithFollowStatus.java
@@ -1,6 +1,6 @@
-package kr.flab.snapnow.domain.user.model.userProfile;
+package kr.flab.snapnow.domain.user.domain.model.userProfile;
 
-import kr.flab.snapnow.domain.user.enums.FollowStatus;
+import kr.flab.snapnow.domain.user.domain.enums.FollowStatus;
 
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
